+++ test/win.test.js
@@ -1759,10 +1759,22 @@ describe("win.js", function () {
   });
 
   describe("::getDefaultShell", function () {
-    it("returns the value of ComSpec", function () {
+    it("returns the value of %COMSPEC%", function () {
       const env = { ComSpec };
       const result = getDefaultShell(env);
       assert.strictEqual(result, ComSpec);
+    });
+
+    it("returns the value of %COMSPEC% when it's the empty string", function () {
+      const env = { ComSpec: "" };
+      const result = getDefaultShell(env);
+      assert.strictEqual(result, "");
+    });
+
+    it("returns 'cmd.exe' if %COMSPEC% is not defined", function () {
+      const env = {};
+      const result = getDefaultShell(env);
+      assert.strictEqual(result, "cmd.exe");
     });
   });
 });